Simplify string-based summary output of cut lists.
This patch modifies the generation of string-based cut list summaries
(as provided by mythcommflag --getcutlist) so that they specify the
beginning and end of recording for cut-to-beginning and cut-to-end cuts
(as was done for previous versions of MythTV).  Since the string
summaries do not specify mark types, this makes parsing easier for
applications that try to use summary data.

Also, don't load placeholder marks in ProgramInfo::QueryCutList().  The
database should never contain placeholders, so no need to load them.

Refs #9389.