Bluray: Disable screengrab/preview generation for bluray discs.
This probably isn't actually being invoked at the moment so this acts as
a guard if nothing else.

Opening a bluray disc can be an extremely resource intensive operation
and under certain situations it can currently lockup. If a still frame
is seen during the initial disc scan, the bluray object will effectively
stall (it keeps returning still frame events and no additional data).
Unlike libdvdnav, there is no mechanism to skip this state in libbluray.

There is work-in-progress to avoid this situation but, regardless,
preview generation is IMHO still a bad idea for bluray.