mythiowrapper: Add callback funtionality to mythfile_open.
This will be used by BDRingBuffer (and possibly DVDRingBuffer) to get
progress updates.

When we open a bluray disc structure, we have one instance of
BDRingBuffer that creates a bluray access object and initialises it.
That initialisation will perform many separate file open calls that are
effectively hidden from the main ringbuffer instance. With potentially
several hundred files to open, BDRingBuffer::OpenFile can block for
several minutes. The callback will be used for UI updates so that the
user receives some form of visual feedback that the frontend is still
working.

N.B. only one callback can currently be registered at a time. This is
not ideal but should suffice for the time being. Any concurrent attempts
to register for callbacks will fail but will only result in no UI
updates.