Bluray: Provide some UI feedback on startup progress.
This uses the mythfile_open callback to update the BDRingBuffer object
every time a file is successfully opened by libbluray during the initial
open/initialisation stage.

The callback then calls processEvents - which will update the
TvPlayWindow widget as main UI drawing is still enabled at this stage.

This relies on the theme using a progress bar widget in TvPlayWindow -
which is pretty much restricted to Graphite at the moment.