MythPlayer: Reorder pause/unpause sequence.
The pause/unpause behaviour will be highly dependant on which thread is
calling Pause(), but broadly speaking the most involved and hence time
critical action is pausing/unpausing the decoder. Hence do this last on
pause so that the video (i.e. main thread) is more responsive and
unpause the decoder first before the video thread is occupied.