Update BlurayMetadata class to split parsing and opening of disc.
Also update those functions to return bool rather than void.

For some stuff I'm working on it's nice to have a convenient function to determine if the disc in the drive is a usable Blu-ray, but it might not necessarily contain Disc Library metadata.  So break it out such that I can behave one way if there's a bluray, and another if there's a bluray with metadata.