Bluray: Use the new kUpdateTvProgressEventType to update TvPlayWindow
This also adds the necessary hookup in the TV class.

TvPlayWindow should now receive a kick each time a new bluray file is
opened. This will update the progress bar (if present) and the call to
qApp->ProcessEvents will repaint the window when the repaint timer has
occurred. Animated images will also be updated if present (which serve
as an alternative progress ticker).

No themes were harmed in the making of this patch - so they'll need
updating to use the functionality.

N.B. The progress bar wraps around when it reaches 100 - this is because
we don't know how much work the bluray object needs to do when opening
the disc structure.