Add code to notify the user when their them has an update available.
Notify a user when their current theme has an update available in
the Theme Chooser.  The Master Backend checks the downloadable themes
index once a day and downloads it if it has been updated.  The
frontends check with the Master Backend to see if their theme has
been updated.  The popup is displayed one time only.  If mythfrontend
is restarted, the popup will be displayed again.  The user can turn
off these notifications via the popup menu in the Theme Chooser.