Pause the RingBuffer before pausing the decoder.
If the RingBuffer has reached, or is very close to, the end of the file
then it may be in a wait state. Pausing the RingBuffer clears this wait
state and hence frees up the decoder thread so that it can be paused
quickly.
(cherry picked from commit c3e1879e0c284b3c22676ee50fb002d4bd994ce7)