Select the Default audio track if the container supports it.
Containers such as MKV permit one to define which audio track is primary, and using this flag means not going through the avformatdecoder guesswork we currently employ.  This is courtesy an anonymous patch on #8900.  I am uncomfortable with applying the part that prefers FLAC to E-AC3/DTS/AC3, though.  There are a lot of people out there who encode secondary tracks to FLAC and do a straight copy of the lossy primary track.  I just don't feel 100% on the idea of preferring it to other more commonplace codecs.

Fixes #8900.