Refactor 'eof' handling in the decoder and player classes.
This moves all eof 'ownership' into the decoder which is then queried by
the player as needed. AvFormatDecoder is extended to reset the picture
buffer eof flag to zero if we are clearing the eof state.

This fixes livetv program transitions where the decoder has already
decided it has hit the end of the file before we switch to the next
program and as a result stops returning frames.

My only concern here is the extra locking now required inside the GetEof
method in MythPlayer. The current plan of action is however to ensure
the decoder locking is 'industrial strength' (Hint: it currently isn't)
before looking at how to minimise the locking required in the main
playback loop.
(cherry picked from commit 679b668e3669a38bd08e93aa1f210ad11da0984e)