Remove some potentially dangerous A/V sync code.
If we are late displaying a frame, under the correct conditions (N.B.
currently impossible) we would display the frame anyway. This is only
safe when we know the display of that frame is not locked to the
vertical refresh. If it is, frame display will block until the next
refresh and we may never get A/V sync back.

Unfortunately we often don't know whether our frame display is locked to
the vertical sync (OpenGL, XVideo) so just drop the frame (which is
currently the only code path used anyway).