Fix menu display on LCD
Patch from #9392

Currently, the main menu is not always updated correctly on the LCD display.
When you enter a submenu, the display is correctly updated to show the submenu.
But when you return to the main menu, the display still shows the submenu until
you change the focus of the elements in the menu. This makes it difficult to
navigate through the menus using only the LCD display.

This commit fixes this behaviour by triggering an updateLCD() when a menu is
aboutToShow().

Signed-off-by: Gavin Hurlbut <ghurlbut@mythtv.org>