OpenGL: Revert 'Don't call QGLContext::doneCurrent() in MythRenderOpenGL'
Originally committed in:-

http://svn.mythtv.org/svn/trunk@27281 7dbf422c-18fa-0310-86e9-

This should never have been committed as I completely ignored the
situation in which the decoder thread re-initialises the videooutput
class when a stream change occurs.

The change will probably break windows OpenGL video playback and from
prior testing, will probably have a negative impact on linux Intel GPU
performance.

Will fix sundry crashes when using OpenGL video playback and the stream
changes (size, interlacing etc).

The proper fix, which I've been mulling for some time, is to make the
decoder object signal to the main thread that the video parameters have
changed and a re-initialisation is needed. This is not a trivial change
though.