Fix seeking after LiveTV transitions.
After a LiveTV transition occured we were setting the max seek to
the amount of frames written by the recorder for the current
segment even though we were still playing a previous segment in the
LiveTV chain. Also, when seeking near the end of a segment we were
setting fftime to -1 which prevented further seeking due to an
incorrect comparison in FastForward.

Regs #9023.