Remove an extraneous QMutex unlock in RingBuffer.
This has been reported several times as the cause of assertions on the
windows build and when using a debug build of Qt. I've not been able to
trigger that condition but logically the current line is the last action
in the destructor of the base class and hence should have no impact.