UPnP: Extend the MythFEXML interface to send actions directly to the frontend.
In summary:-

Send a message to the frontend (N.B. syntax change).
- http://mythfrontend-ip:6547/MythFE/SendMessage?text='Hello world'

Send an action (e.g. UP) to the frontend.
- http://mythfrontend-ip:6547/MythFE/SendAction?action=UP

A list of valid actions is cached internally and used to validate the
submitted action. To retrieve a simple XML formatted list of those
actions, their descriptions and the contexts in which they are valid.
- http://mythfrontend-ip:6547/MythFE/GetActionList

To display an html test page of all valid actions.
- http://mythfrontend-ip:6547/MythFE/GetActionTest

GetServeDesc returns the service description for the first time (the
description needs updating).

N.B. There is every possibility that this interface will be tweaked
before 0.25 is released.

There are some issues around languages/translations with the action
descriptions and I make no apologies for the quality of my html/xml
(patches willingly accepted)