Convert DeviceReadBuffer from pthreads to QThread
As part of ticket #5501, I'm chasing down all direct users of pthreads and
converting them to QThreads.  This victim is the thread within the DevRdB code.
This chunk of code could use a refactor, likely, but I have attempted to keep
it as close to the previous code as possible.

This code has been running in production here for 12 days.