Change to using QPointer so double delete issues stop
We keep on seem to get some race conditions with myth_system launched things
exiting too fast (i.e. command not found, fast crashes) and causing a set of
race conditions involving deleting the MythSystemUnix item when the process
ends.  The use of QPointer should avoid these problems as when a QPointer item
is deleted, the reference is automatically zeroed, even if it was deleted
elsewhere (or so the docs seem to say).  With that plus a few more null checks
and we can avoid using it after it's freed.