Store calculated duration in the db
At recording end, a value of recendts - recstartts (converted to ms) is stored
in the recordedmarkup table.  At the end of commflag or a seektable rebuild,
the accurate number of ms (as determined by adding frame times during playback)
will overwrite this value.

On playback, this value is checked first.  If the value is in the database, it
is used for the duration in the OSD.  If it is not there, then the duration
from ffmpeg will be used instead.

To get this working, it was necessary to reset the calculated duration after
the logo scan in commflagging, otherwise the end result was incorrect.