Keep 10 days of old program listings.
Keep program listings as long as we keep matching, non-recorded
recording history.  Before, we were keeping 7 days of listings and 10
days of recording history.  This is more consistent and makes
post-mortem analysis easier (allowing users to see what shows aired at
the time of their missed recording for as long as we keep the missed
recording information).