Make it so the second frontend won't crash
If for some reason, a user decides to run two frontends on the same box, and
forget to set the service port to something non-standard for the second one,
we want it to at least not crash.

With this commit, it will just give up on running a webserver on the second run
frontend, but it will run.

Several tweaks were necessary to make the deleted webserver not double-free
and not otherwise corrupt memory.