Improved handling of 'Show the Program Guide when Start Live TV'.
There is no longer a need to hit the database for the appropriate action
and add a simulated keypress to the list of queued keypresses (and hope
the player is up and running before the keypress is processed).

Instead just start the guide when the player is ready.

Cherry plucked from 341efdf947f4c19558d1