Don't check whether there are pending events in the TV playback loop.
Initial feedback suggests this is the cause of recent playback issues on
certain OS X builds. It appears that on some Qt builds hasPendingEvents
is always returning false and hence we never return to the main event
loop. The OS then believes the app is unresponsive to external events
and throws up the Spinning Beach Ball of Death.