Add a simple http/html 'remote control' to the frontend.
To test, point your browser at http://frontend-ip:6547/MythFE/GetRemote

This is probably nothing more than a proof of concept at the moment and
has certain limitations around translations and responsiveness on
smartphone interfaces (well, on my iPhone anyway!).

It should hopefully prove useful as an inspiration for others to get
busy with some more interesting versions and as a test interface into
the frontend for those trying to debug lirc or other input issues.