Fix passing a URL key/value containing %26 (escaped &)
It seems that this code *only* deals with Parameters in the URL (maybe on POST
too), but not in SOAP requests from UPnP clients.  It should be a benign
change to not mess with %26 in the params string before splitting on &.  Any
self-respecting client will know that & separates the keys, and not to escape
them.  If this breaks any specific clients, we will have to put in a workaround
for the broken client perhaps.

Fixes #8132.