Auto-detect UI painter.
If running on Windows, use D3D9.  Else, try the OpenGL UI painter.  If
initialization fails or if OpenGL is not using direct rendering, fall
back to the Qt painter.  This also removes the ThemePainter setting.

The auto-detection will automatically choose the Qt painter for ssh X
forwarding and such, but will use OpenGL painter for local display (when
OpenGL is properly installed and configured on the underlying system) so
users shouldn't need to worry about overriding the painter.

Theme developers may test with a non-default painter by starting
with -O UIPainter=[opengl|qt|d3d9].