Remove util.cpp from libmyth.pro
The file was moved into libmythdb (now libmythbase) in f986c9b163e242, but
it wasn't removed from the .pro file.   This caused that source file to get
compiled into both libraries, which caused an issue in #9586 after the split
of the mythexp.h in 26ea6674c8fd5

Distclean is suggested.