Try to keep from another apparent race condition
Not quite sure why the QMap is going crazy, but let's try reading the
AutoCleanup flag before unlocking so we can mitigate the race some.  Also,
in both myth_system() and in the signal handler, use deleteLater() so we don't
try to double-delete, as that will just cause other messes.

Fixes #9602