Improve timestamp logging and restrict fixup code.
The new logging is greatly improved to provide information such as
which timestamps are active (reordered or dts) and if fixups are
being invoked.

We now only provide fixups for timestamps that are equal to or upto
one frame interval less than the last video timestamp. The previous
code would fixup timestamps that were upto 10 seconds behind the
last used video timestamp. This caused the code to start tracking a
bogus timestamp upto 10 seconds in the future. I'm seriously
considering just removing fixups for timestamps in the past. We now
use reordered pts timestamps and if they have issues then we switch
back to using dts. This should mean that we really only need fixups
for missing timestamps which are normal for some types of video. It
would be interesting to know if there are any videos out there
without any good type of timestamps available and if so what we
could realistically do about it.

Backport from 9af38c8b3d7091c4bd84 since it provides a much easier
way to debug timestamp issues until the next release.