Another run at getting rid of symbol issues
Change version.cpp -> libs/libmythbase/version.h, which we then include in
mythversion.h.  Also convert from extern const char * -> #define constants.
This will remove all of the symbols.  However, it comes at the cost of a
slightly slower re-compile as every file that uses the constants will need
recompiling.  If we wanted to get around that, we can use functions in
libmythbase as the access points, such that only one file needs a recompile,
and it is buried in a library, but still anything using said library will need
relinking.

Can't win, really.  Refs #9586