Convert backend threads to QThreads
Pursuant to #5501, I have converted a pile of threads in the backend to QThread
from pthreads.  These include: Scheduler, Expire, Update, HouseKeeper,
FillDB, Delete, Truncate.  This has been running under load on my production
backend for most of a week, and seems to be fully operational.

Make distclean is suggested, but likely not required.