MythUIWebBrowser: allow unsupported content to be downloaded or played.
This updates the web browser widget to ask if you want to download any files
that the browser doesn't know how to handle. You can also right click on an
image and download those as well. This is primarily for the music metadata
editor so you can do a search for a cover image for MythMusic to use.

Any audio files that are clicked and MythMusic can play will also ask if you
want to play them in the music player. Currently mp3 files play well but other
formats no so well but I do intend to look at improving that. There is support
to also play video files but this is currently disabled until the internal
player can better handle playing from an http url.