Hardware Profiler: Update opt-in hardware profile monthly.
Checks for opt-in/existing profile and last date of update.  If this was greater than a month ago, update the existing profile.  Seems to work nicely in my testing so far.

Plans are afoot to have the housekeeper update the hardware profile for backend-only machines, the code should be very simple.