Make NULL always fail.
Change of behaviour with the NULL device. It will now simply fail when trying to open it. The consequence of this is that audio will be disabled which was ultimately why the NULL device exists to start with.
Change the NULL device to report that it supports all audio types, channels and format, so audio processing will never occur