Don't use QDateTime::toString() for prepared queries.
When using bindValue() to specify datetime values, do not use
toString().  Instead, pass the QDateTime to bindValue() so that it can
pass the value using the MySQL-C-API MYSQL_TIME type--so that database
datetime string literal format isn't an issue.  Refs #8585.

Thanks, again, to Rob Smith for finding another incorrect pattern to
search for.