Allow near-end bookmarks for in-progress recordings.
When exiting in-progress recordings, what's now near the end may not be
near the end later, so go ahead and set a bookmark (for those who have
specified "Save position and exit"), so users don't lose their place.

I didn't modify MythPlayer::SetWatched(), since I have a feeling quite a
few users will watch in-progress recordings and exit after the end of
the show, but before recording stops--in which case they have watched it
and just want to be done with it (rather than have to manually mark it
watched).  On the other hand, users who watch until they're caught up
with current time then exit before the end of the show will likely come
back to finish watching the recording sooner, rather than later (so it
will be watched soon enough).  Alternatively, if the auto-detection does
the wrong thing, they can choose to mark the show unwatched.

Since we can't always do the right thing automatically, I'm trying to
guess what's likely right for the most people.  That said, I have some
projects in the works to improve both bookmark and watched handling.

Fixes #9625.