Windows: Fix a crash in DXVA2Decoder when codec not supported.
Ensure our decoder surfaces are set to NULL and fix a crash when expected codec support is not available.

This avoids the crash but video playback will fail as we can't easily fallback to a different (ie. unaccelerated) codec - we need a full codec support check but it's not obvious how that can be done easily.