Add a couple of jump points to show each widgets bounding rect and name.
This is really only useful for developers or theme creators to help see
exactly the area a widget is occupying. It makes it easier to set out text
areas so they don't overlap other widgets for example.

The jump points are 'Toggle Show Widget Borders' and 'Toggle Show Widget Names'
and both are unbound by default.

The show names jump point will show the widgets name in the top left of the
widgets bounding box but only when show borders is turned on.