Fix jobqueue thread shutdown
Originally, this had a pthread_testcancel in the thread, and pthread_cancel in
the controlling code.  I have changed this to get the same sort of controlled
thread shutdown that it used to have by putting in a flag for "processQueue".
When you want to exit that thread, tell it not to process anymore, and next
time the thread loops, it will exit (just like before with the testcancel at
the head of the loop).