Change ALSA buffer size and period size
With some hardware, the previously used value (200ms/50ms) led to buffer underruns. We increase the buffer size to 0.5s and the period to be 1/16th. This push the minimum amount of allocated RAM to ALSA hardware buffer to 128kB (default is 64kB). This will lead to extra verbosity in the log, but if it was working with 200ms then those warnings/errors can be ignored