OpenGL: Use GLSL to draw rounded rectangles.
N.B. The GLSL/OpenGL 2.0 profile is not enabled by default, so you won't
see the glorious benefit of this commit :)

- Add new default shaders to draw circles, circle edges and horizontal
and vertical lines. All implement some simple but effective anti-
the curved and straight edges - without them the results are
disappointing.

- Add a new generic parameter matrix for passing parameters to shaders.

- Use the new shaders to render rounded rects that DON'T use a gradient
when using the OpenGL 2.0 profile. The rendering code looks complicated
but is actually nothing more passing a series of vertices and parameters
to the GPU.

This still needs a some tweaking to better deal with thick edges. At the
moment they are too 'fuzzy' but 1 and 2 point lines should be fine.

The existing line drawing for the basic (unrounded) rect also needs
converting to the new line shaders. Without them they are actually
rendered in the wrong place.