MythNetvision: Use the backend downloader to download saveable files.
Pop up a progress dialog and track the download as the backend performs the download into the Default Storage Group.  Offers an experience identical to the download done in the simple video setup wizard.

This is far more useful that the traditional behavior where the file was downloaded in the background to ~/.mythtv and nothing much was done with it.  It's not 100% useful yet as I'm still missing the step to insert a recorded item a la mirobridge, but that's the only bit of this remaining.

TODO:

1) Add a similar download behavior to the video search.
2) Insert downloaded videos as recordings.
3) Update mythfillnetvision to download and insert new RSS articles where the user has configured automatic download of videos.