Make mythcommflag terminate early on 0-byte files
I had a rash of 0-byte files the other day, and I found that mythcommflag
would sit and spin trying to create some initial buffers, or something along
those lines.  This change makes it check if the file is 0 bytes.  If it is,
it will immediately quit with an error, rather than indefinitely try to
grind on it.

This may need some tweaking to deal with commflagging while recording as it
could cause a race condition if the file has no data at the moment that
mythcommflag starts.