Some fadvise changes.
First this gets rid of DONTNEED advice. These calls a global, so in order for these to be safe in the MythTV context we'd need to make sure that no other process is currently accessing the file. This is complicated on 64 bit platforms where you can memory map the whole file and extreemely complicated on 32 bit platforms.. Instead lets leave this task to the OS.

Second we change some of the WILLNEED calls. On Linux these are unfortunately implemented as blocking calls; that is they block until the requested number of bytes have been read, instead of just telling the OS to start reading in those bytes in a kernel thread. Consequently, this drops the calls in some cases and reduces the number of bytes to read in other cases.