Fix commflag bailing on remote files
Fixes #9663

Deals with remote files on commflag properly after adding the check for 0-byte
files.  Now it will check the remote file for existance and non-0 length,
or the local file for existance and non-0 length.  Once again, if this causes
issues with during-recording commflagging, we can add some delay and recheck
or something like that to avoid race conditions if necessary.