MediaMonitor::GetMedias: Search for multiple media types
Add the ability to request multiple media types when searching MediaMonitor
disks.

This is useful in the Gallery & MythGallery plugin which registers specific image
filetypes with MediaMonitor.  When a CD/USB is inserted with media
containing images, the disk is searched statistically and the media type
is set to MEDIATYPE_MGALLERY.  However, when the galleries show a list of
devices with supported media it needs to ask for both MEDIATYPE_MGALLERY
and MEDIATYPE_DATA.
The Galleries are the only user of the MediaMonitor::GetMedias function.

Refs #9520

Signed-off-by: Roger Siddons <rsiddons@mythtv.org>