libmythui: Fix fullscreen windowing with Compiz (and Gnome?)
This is based on the patch from Lawrence Rust in ticket #9589 which
notes that Compiz requires the window to be visible before we set its
state.

Testing across various Ubuntu desktops (with and without Compiz and
composite enabled) and on OSX and windows suggests that it is safe just
to show the window earlier in the initialisation process and avoid the
ifdef.