Check MYTHHTMLDIR env var for static html server doc root for testing.
If the MYTHHTMLDIR environment variable is set, use the specified
directory for the doc root of the static html server in the backend.
This is primarily to aid in testing as we work on the setup rewrite,
a user should not normally change the doc root since that would break
the web-based setup.