Add a simple Storage Group Directory editor to the backend web setup.
Users can add directories to and remove directories from existing
Storage Groups.  There isn't a way to add a new Storage Group yet.
All Storage Groups for all hosts may be manipulated and there is
a checkbox to allow just viewing/editting the current host's
Storage Groups.

This is a work in progress, and needs a few more features, but
it's a start...