Initialise audioOutput object if it exists.
Code incorrectly assumed that the only place the creation of the audioOutput object was done in MythPlayer. Mythtranscode creates its own audioOutput object and use AudioPlayer::SetAudioOutput to set it.
Now if an audio object exists, we initialise it no matter what. This will likely increase verbosity in some cases

Fixes #9425