Internet Content: Don't use the chromeless Youtube player.
Display the "OSD" for the youtube player to be consistent with the other fullscreen players.  It's not ideal as every info OSD for each player is different.  While most can be suppressed entirely, the perfect solution would be the ability to init a MythUI OSD object to pop up over the browser.  I can't see a way to do that, though, as the current OSD isn't installed for plugins and relies on a MythPlayer.  It would be nice to be able to create a more generic OSD that I could feed curtime/totaltime/title/description to, but I can't see an obvious way to do that right now, so in the meanwhile, just display some of this information in its native context.