Don't check for existing theme info cache in housekeeper.
Instead of checking for the theme info cache .zip file in the housekeeper,
just run the download job once a day. This prevents a user's log being
spammed with download errors if they don't have internet connectivity
on the master backend or if they are running a non-master, non-released
branches of MythTV.  Currently non-master, non-released branches try
to download a 0.25 themes.zip index which doesn't exist since MythTV 0.25
doesn't exist yet.  That issue will be dealt with in a future patch.