libmythtv: Enable visualisations.
- libfftw3 support is required at compile time and MUST be forced with
--enable-libfftw3 (existing libfftw3 support in configure appears to be
a hangover from previous code, it isn't actually used for anything else)

- much of the code is taken from mythmusic.

- visualisations will only work with 2 track audio streams.

- the visualisation is rendered over any video (and Pip) but beneath the
OSD.

- 2 visualisations are currently available. The 'default' is a simple
implementation of Spectrum which uses libmythui rendering routines and
is available when using the VDPAU, OpenGL and Direct3D9 renderers(i.e.
those with accelerated, overlaid OSDs). Circles is available when using
OpenGL 2.0 (not currently enabled) and, while a little easier on the
eye, is more a proof of concept to ensure correct inheritance,
dependency checking etc is working. It doesn't actually rely on the
features available with the OpenGL 2.0 renderer - though drawing of
circles is accelerated with that renderer.

- there is no technical reason why either visualisation will not work
with software rendered OSDs (e.g. XVideo). The limitation is purely a
performance consideration.

- adds a new keybinding "TOGGLEVISUALISATION" which is unbound by
default, otherwise the visualisation can be toggled via the new entry in
the OSD menu or the http interface.

- it should be easy enough to port any other mythmusic visualisations,
though libvisual integration is more difficult (and may not be
possible). If additional visualisations are made available, I'll add a
mechanism to chose which to use.