Backend setup webpage tweaks.
Minor layout changes to create a 'setup wizard' placeholder and
put the other setup pages under and 'advanced' section.

Add a 'host' dropdown on top of the Storage Group editor page
which allows editing the Storage Group list for any host
individually.  A 'Display All Hosts' option is present at the
top of the list to allow viewing Storage Groups for all hosts
as a whole.

Add a placeholder button on the Storage Group editor screen for
a 'Browse' option for picking a new directory.

Add a placeholder link on the Storage Group editor screen for
deleting a whole Storage Group without having to delete each
directory in the Storage Group individually.

Minor CSS cleanup.

Add some 'back' type links on screens to allow going back to
the previous screen.