Update binary API version after mythtvsetup-www merge.
Make clean, etc..

NOTE: A lot of the code in the mythtvsetup-www branch came from Daniel's
patch in ticket 8262 (http://code.mythtv.org/trac/ticket/8262) and there
is still a lot of work to be done on that patch.  This branch was merge
into master to make it easier for developers to work on the backend web
setup.

As noted in a previous commit, the setup webpage is currently
disabled by default until we have some form of authentication working.
To enable the backend /setup/* and /Config/* web pages, you must
set an environment variable called MYTHHTMLSETUP.  The value of
this variable does not matter, the code only checks for its
existence.  If you're using a shell like bash, the easiest way to
test is to run mythbackend like this:

    MYTHHTMLSETUP=1 mythbackend