Minor setup page tweaks/examples.
Put a warning at the top of the setup page to make sure that
users and devs realize this is a work in progress and may be
non-functional in some areas.

Add some working example links showing how we can jump out of
the wizard into an advanced setup page if the wizard page does
not provide all the settings the user needs access to.