HTML Setup: Use CSS to hide the DirID values in the SG editor.
I have not confirmed that this doesn't break the actual remove SG Javascript function, but by my reading, it should still work fine.  Hiding this to make the whole approach more friendly, but done without group consultation so I will live if there's some reason we need to display this.