HTML Setup/API: Add a method to test DB Settings.
Takes username, pass, db name, db host, and port and returns a bool true or false on whether a DB connection was possible.  Simultaneously, hack in a button on the DB setup page of the HTML Setup wizard to test the DB connection.  Again, this should probably end up an AJAX-y sort of thing that is validated as the text is input, or on widget change, or whatever, but right now it's just a popup.