mythplayer: Revert back to the old method used for duration/position in the OSD slider.
This has proven to be problematic given that many encoders produce bad timestamps or
reset the timestamp (timestamp discontinuities) yielding this method insufficient. In
trunk Gavin has added the ability for commflag to generate and store the duration.
Using the commflag duration will allow accurate duration/position for broken videos.

Refs #9109.