A bunch of HTML Setup changes.
Add an edit 'popup' div for editting things outside the content div.

Add a JobQueue Job editor to allow editting JobQueue Job commands and
descriptions.  The backend settings such as the window, which jobs can
run on a host, etc. are displayed, but not saveable yet.

Modify the Setup Wizard so that it displays in the 'popup' div.

Add status and error message areas at the top of the page in the black band.

Convert the Storage Groups editor to use the status message area at the top
of the page instead of using alert().

Convert the database test code to use the status message area at the top
of the edit popup.

Convert the password changer to use the status message area at the top of
the page.

Modify HttpConfig, MythSettings, and MythSettingsGroup to allow editting
a single group of items instead of the whole list in the XML file.  This
is used in the JobQueue backend settings tab and will be used in other
places as well.

More code to get toward the saving of settings in HttpConfig.  The new values
aren't validated or saved to the DB yet, but the backend does return a
JSON list of key/value pairs which will contain status messages about any
values that do not fit the criteria specified in the XML, such as an
integer out of range.  This status message is displayed in a div by the
setting input on the form.  This will give us error messages in context.
Eventually there should be no errors returned because we hope to catch these
in javascript prior to submitting.

Split up the 'General' page into multiple tabs. This is based on the groupings
in config_backend_general.xml and is just for readability, the groupings will
probably change.

Minor CSS cleanup to display some 'content' div data a little better.