HTML Setup: Server Side Scripting sample changes.
Rework the menu a little to allow the sample pages to work like
the rest of the site rather than having them have their own menu
page, etc..

Comment out the '<img src' tag on the recorded.qsp sample.  This
can take a long long time if you have a lot of recordings.  Leave
the code in as an example.

Add column headers to recorded.qsp.

Add a "Title - SubTitle" column to recorded.qsp.