HTML Setup: Remove new jQuery.ajaxSetup() call in /index.html
This call was copied from samples/index.qsp before removing that
file, but these 'default' settings break loading of other pages
since we load a lot of content into divs using jQuery.  Specifically,
on screens with tabbed content, the tabs are all empty with these
settings.  The samples pages still appear to work, so I'm removing
this call for now.