HTML Setup: Wizard enhancements, data loading, data validation, etc.
Added some placeholders to the wizard.

Started hooking up data validation.  The only thing hooked up right
now is integer_range and only for the wizard, but the basics are there
to start validating in other places either real-time or upon saving.

Supporting code added to HttpConfig and mythsettings.cpp to allow
getting the raw settings list that goes along with the config XML
that we can already download.  This allows us to prepopulate the value
for inputs which are not built using HttpConfig such as the wizard.

Some other minor fixes along the way, <h2> -> <div class='Title'> in
some places, etc..