HTML Setup: Directory Browser, Misc. fixes, Minor cleanup.
Add a popup directory browser for adding Storage Group directories.  This is
available in the Setup Wizard and the Advanced Storage Groups Editor.  This
pulls in a copy of jqueryFileTree which is GPL-ed.  License info is at the
top of jqueryFileTree.js.  This is a slightly modified copy of jqueryFileTree.
I have added the ability to highlight a directory for saving it as well as
adding a 'dirsOnly' argument to allow browsing only directories which fits
our use case.  The browser will get some more love, but it is functional
currently.

Clear error messages by form inputs in the advanced editor when submit is
clicked.  These will be reset if the save fails (when we get saving hooked
up on the backend).

Fix a race condition in the Setup Wizard startup where the setting values
would not be available yet when the tabs were drawn and populated.

Bugfix in mythsettings.cpp fill_setting() by natanojl on IRC. Previous code
tried to access an item out of bounds.