HTML Setup: Settings help, System Events, Sized text inputs.
Add a popup help box for each setting which contains the value of the
help_text from the config_settings_*.xml file.  This link is available in
the settings editor and in the advanced editor.

Add system events to config_settings_general.xml and make these
settings show up in the system event advanced page.  Saving and validation
are still not hooked up since these use the same /Config/General* code
as the other settings.

Add some code to set a few custom sizes for the different setting
types that use a text input.  This way they are not all 20 characters
wide.  String settings use a longer input size than integers, etc..